+++ src/background/main.rs
@@ -25,10 +25,10 @@ use std::io::{BufWriter, Write};
 use color_eyre::owo_colors::OwoColorize;
 use error_handler::Result;
 use exposed::register_invoke_handler;
-use modules::cli::{
+use modules::{cli::{
     application::{handle_cli_info, SEELEN_COMMAND_LINE},
     infrastructure::Client,
-};
+}, tray::application::force_tray_overflow_creation};
 use plugins::register_plugins;
 use seelen::SEELEN;
 
@@ -97,6 +97,7 @@ fn main() -> Result<()> {
         .setup(move |app| {
             log::info!("───────────────────── Starting Seelen ─────────────────────");
             Client::listen_tcp()?;
+            force_tray_overflow_creation()?;
 
             let mut seelen = unsafe { SEELEN.make_guard_unchecked() };
             seelen.init(app.handle().clone())?;