+++ src/background/modules/mod.rs
@@ -1,5 +1,5 @@
-pub mod power;
-pub mod file_extensions;
 pub mod cli;
+pub mod file_extensions;
 pub mod input;
+pub mod power;
 pub mod tray;