+++ src/background/modules/tray/mod.rs
@@ -1,3 +1,3 @@
-pub mod domain;
 pub mod application;
+pub mod domain;
 pub mod infrastructure;