+++ src/background/seelen_shell/mod.rs
@@ -10,4 +10,3 @@ impl SeelenShell {
         Self { handle }
     }
 }
-